+++ src/sptensor/add_omp.c
@@ -3,7 +3,7 @@
 
 
 /* TODO: bug. */
-int sptSparseTensorAddOMP(sptSparseTensor *Y, sptSparseTensor *X, int const nthreads) {
+int sptSparseTensorAddOMP(sptSparseTensor *Y, sptSparseTensor *X, size_t const nthreads) {
     /* Ensure X and Y are in same shape */
     if(Y->nmodes != X->nmodes) {
         return -1;
@@ -25,12 +25,12 @@ int sptSparseTensorAddOMP(sptSparseTensor *Y, sptSparseTensor *X, int const nthr
     free(dist_nrows_Y);
 
     printf("dist_nnzs_Y:\n");
-    for(int i=0; i<nthreads; ++i) {
+    for(size_t i=0; i<nthreads; ++i) {
         printf("%zu ", dist_nnzs_Y[i]);
     }
     printf("\n");
     printf("dist_nnzs_X:\n");
-    for(int i=0; i<nthreads; ++i) {
+    for(size_t i=0; i<nthreads; ++i) {
         printf("%zu ", dist_nnzs_X[i]);
     }
     printf("\n");
